Addressed silent issue with fasta definitions
Encountered a silent issue where the tiles for chromosome U where
generated with wrong chromsome name for chromosome U ('>chru   md5')
because it had an invalid definition line where a semicolon was missing
between two columns. We are now throwing an exception in those cases to
alarm the user that the input data has a problem.