Removed chrom_name from definiton fields
The first string of any fasta definition line inside a genomic reference
file is the chromosome name prefixed by a '>'. Unfortunately this
doesn't get delimited by a ';' and screws up the parsing of the column
variables (i.e. '>chry type' instead of 'type'). From now on we remove
this part of the string before splitting the definition line into
fields.

Also moved call to GenomeTiler#definition_to_fields out of a loop
because the definition string doesn't need to be split for every window.